+++ src/organizer/qorganizercollection.cpp
@@ -220,9 +220,9 @@ QVariant QOrganizerCollection::metaData(MetaDataKey key) const
     \relates QOrganizerCollection
     Returns the hash value for \a key.
  */
-Q_ORGANIZER_EXPORT uint qHash(const QOrganizerCollection &key)
+Q_ORGANIZER_EXPORT size_t qHash(const QOrganizerCollection &key)
 {
-    uint hash = qHash(key.id());
+    size_t hash = qHash(key.id());
     QMap<QOrganizerCollection::MetaDataKey, QVariant>::const_iterator i = key.d->m_metaData.constBegin();
     while (i != key.d->m_metaData.constEnd()) {
         if (i.key() == QOrganizerCollection::KeyExtended) {
